Refactor Body to use generic vector types
Updated the Body struct and related methods to use particular's VectorDescriptor for improved generic compatibility. Adjusted position and force handling to work with [f32; 2] arrays instead of glam::Vec2, and updated the physics step logic accordingly. Also set wgpu surface alpha_mode to Auto for improved rendering configuration.

+++ src/render/frame_descriptor.rs
@@ -26,7 +26,7 @@ impl FrameDescriptor {
         for body in sim.physics_context.bodies.iter() {
             transforms.push(GpuTransform {
                 model: Mat4::from_scale_rotation_translation(
-                    Vec2::splat(2.0 * body.radius()).extend(1.0),
+                    Vec2::splat(2.0 * body.radius() as f32).extend(1.0),
                     Quat::from_rotation_z(body.rotation()),
                     body.position().extend(1.0),
                 )
